requirements.txt
pillow
pillow_heif
pyheif